Change return values of functions for consistency
By default, all functions will return 0 on success, and a negative value
on failure.

Functions checking a window state (functions named `wm_is_*`) will
return 1 on success and 0 on failure. This is for semantic reasons:

	if (wm_is_mapped(wid))
		// do stuff

Some functions have a return value that is not representative of the
success of said function, and instead will return an item specific to
their usage.