Allow setting window border as ARGB values
Border colors are represented as 32 bits integers, out of which the 2
first byte reprensent the alpha channel (in TrueColor mode, which
we assume).
When the very first bit is set (eg. 0x80123456), the signed integer
representation would be a negative number.
If we want to handle the alpha channel correctly, we cannot check the
color value as an integer, and accept any value.

The obvious advantage is that setting transparency now works as
expected. The drawback being that if you only want to change the border
size, you still have to set the border color. Deal with it :)